Update for environmental awareness: - List saved places in !stats Code cleanup
