+++ src/agent/npc/utils.js
@@ -118,6 +118,74 @@ export function itemSatisfied(bot, item, quantity=1) {
 }
 
 
+export const pad = (str) => {
+  return '\n' + str + '\n';
+}
+
+
+export function getAdjacentBlocksString(bot) {
+  let result = '';
+  let below = bot.blockAt(bot.entity.position.offset(0, -1, 0));
+  if (below?.name && below.name !== "air") {
+      result += `\nStanding on: ${below.name}`;
+  }
+  let standingIn = bot.blockAt(bot.entity.position);
+  if (standingIn?.name && standingIn.name !== "air") {
+      result += `\nStanding in: ${standingIn.name}`;
+  }
+  let headIn = bot.blockAt(bot.entity.position.offset(0, 1, 0));
+  if (headIn?.name && headIn.name !== "air") {
+      result += `\nHead in: ${headIn.name}`;
+  }
+  let above = null;
+  let maxAboveHeight = 10;
+  let currentAbovePosition = 1;
+  while ((!above?.name || above?.name === "air") && currentAbovePosition < maxAboveHeight){
+      currentAbovePosition += 1;
+      above = bot.blockAt(bot.entity.position.offset(0, currentAbovePosition, 0));
+  }
+  if (above?.name && above.name !== "air") {
+      result += `\n${currentAbovePosition} block(s) above you: ${above.name}`;
+  }
+  return result;
+}
+
+
+export function getInventoryString(agent) {
+  let bot = agent.bot;
+  let inventory = world.getInventoryCounts(bot);
+  let res = 'INVENTORY';
+  for (const item in inventory) {
+      if (inventory[item] && inventory[item] > 0)
+          res += `\n- ${item}: ${inventory[item]}`;
+  }
+  if (res === 'INVENTORY') {
+      res += ': Nothing';
+  }
+  else if (agent.bot.game.gameMode === 'creative') {
+      res += '\n(You have infinite items in creative mode. You do not need to gather resources!!)';
+  }
+
+  let helmet = bot.inventory.slots[5];
+  let chestplate = bot.inventory.slots[6];
+  let leggings = bot.inventory.slots[7];
+  let boots = bot.inventory.slots[8];
+  res += '\nWEARING: ';
+  if (helmet)
+      res += `\nHead: ${helmet.name}`;
+  if (chestplate)
+      res += `\nTorso: ${chestplate.name}`;
+  if (leggings)
+      res += `\nLegs: ${leggings.name}`;
+  if (boots)
+      res += `\nFeet: ${boots.name}`;
+  if (!helmet && !chestplate && !leggings && !boots)
+      res += 'Nothing';
+
+  return pad(res);
+}
+
+
 export function rotateXZ(x, z, orientation, sizex, sizez) {
     if (orientation === 0) return [x, z];
     if (orientation === 1) return [z, sizex-x-1];
